make 3 helper files to cleanup main file + remove unncessary files
